Build Windows ARM64/aarch64 wheel
On Windows ARM, the user must build the `regex` wheel locally,
which requires installation of compiler tools. This PR instead
tells cibuildwheel to build a Windows ARM64 wheel in addition
to the already handled `x86` and `AMD64` targets.

This was tested in Github Actions and produced a working wheel,
installable without local build on Windows 11 Pro ARM64. I tested
that it will import in Python, but did not extensively test the
module other than that.